api/requirements.txt
click==8.1.3
Flask==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pyzmq==23.2.0
Werkzeug==2.1.2
